PICS-2080
Refactored part of the audit question html structure / css.

The view and the edit feature of the audit should be using similar html and css.  It was refactored from using duplicated code.

The JS had the $.loads rewritten to use $.post. It was also cleaned up for saving, reseting, and verifying.

New star was added for flavor.

Checkboxes were integrated "inline" with questions.  Slight modification for questions that are required.