PICS-4962 Pulling business logic out of SqlBuilder.
After moving buildSql() into SqlBuilder.java, I noticed that the functions
setBaseModelFromReport() and initializeSql() were always being called in
order, so I decided to combine them. Now that we're being passed a BaseModel,
there's no need to store baseModel in the class.

The intent of SqlBuilder is this: you give it report parameters for your report,
and it builds the SQL to query everything you need and hands it to you. It
should not have state stored in it, that's why I'm removing baseModel. I've
also added a method, getBaseModel(), to Report.java, because the Report class
should ideally do a lot of the heavy lifting for Dynamic Reports.

In SqlBuilderTest.java, I just always pass in the BaseModel to initializeSql().