PICS-6320 Added link from Manage Audit Type and Category to pull all translation keys
The link will redirect the user to Manage Translations while putting
in the tracer cache all the questions and subcategories of that level.
Added a method in Entity Factory for creating an Audit Type.  Created a class
called TranslationsKeyGenerator that will iterate through a category or
audit type to generate the keys.  Put this class under test.