PICS-12854 : Add liquibase maven plugin to pics-database submodule.
Liquibase can now be invoked by  while inside the
pics-database submodule. The module will not be packaged with the liquibase
files or plugin -- the plugin is available from source code only.
It will look for liquibase significant files in pics-database/liquibase/ .
If we want to make liquibase update part of the deployment process, uncomment
the execution goals inside the pics-database pom.xml .